Add `LottieView` to `MuseView` & add `MuseViewModel` in separate file
Screen Displays a `LottieView` when there are no songs to play
